requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.27.1
python-dotenv==0.19.2
supabase==0.7.1
virustotal-python==0.1.3
python-nmap==0.7.1
urllib3==1.26.9
werkzeug==2.0.3
pyopenssl==22.0.0
reportlab==3.6.12
pillow==9.2.0
